+++ src/professions/admin/check-your-answers.controller.ts
@@ -81,6 +81,8 @@ export class CheckYourAnswersController {
       additionalOrganisation: draftProfession.additionalOrganisation
         ? draftProfession.additionalOrganisation.name
         : null,
+      registrationRequirements: version.registrationRequirements,
+      registrationUrl: version.registrationUrl,
       mandatoryRegistration: version.mandatoryRegistration,
       regulationSummary: version.description,
       reservedActivities: version.reservedActivities,