+++ src/professions/admin/interfaces/check-your-answers.template.ts
@@ -10,6 +10,8 @@ export interface CheckYourAnswersTemplate {
   organisation: string;
   additionalOrganisation: string;
   mandatoryRegistration: string;
+  registrationRequirements: string;
+  registrationUrl: string;
   regulationSummary: string;
   reservedActivities: string;
   protectedTitles: string;