+++ src/professions/admin/interfaces/registration.template.ts
@@ -1,0 +1,7 @@
+import { RadioButtonArgs } from '../../../common/interfaces/radio-button-args.interface';
+export interface RegistrationTemplate {
+  mandatoryRegistrationRadioButtonArgs: RadioButtonArgs[];
+  change: boolean;
+  captionText: string;
+  errors: object | undefined;
+}